Move ShipmentInfoList into DefinitionLists folder
