+++ src/components/Office/ShipmentApprovalPreview.jsx
@@ -11,7 +11,7 @@ import { mtoShipmentTypes } from 'constants/shipments';
 import AllowancesList from 'components/Office/DefinitionLists/AllowancesList';
 import CustomerInfoList from 'components/Office/DefinitionLists/CustomerInfoList';
 import ShipmentContainer from 'components/Office/ShipmentContainer';
-import ShipmentInfoList from 'components/Office/ShipmentInfoList/ShipmentInfoList';
+import ShipmentInfoList from 'components/Office/DefinitionLists/ShipmentInfoList';
 import ShipmentServiceItemsTable from 'components/Office/ShipmentServiceItemsTable/ShipmentServiceItemsTable';
 import { Modal, ModalContainer, Overlay } from 'components/MigratedModal/MigratedModal';
 import { MTOShipmentShape, OrdersInfoShape } from 'types/order';
